+++ src/TronNet/TronECKey.cs
@@ -44,7 +44,7 @@ namespace TronNet
 
         internal byte GetPublicAddressPrefix()
         {
-            return _network == TronNetwork.MainNet ? 0x41 : 0xa0;
+            return (byte)(_network == TronNetwork.MainNet ? 0x41 : 0xa0);
         }
 
         public static string GetPublicAddress(string privateKey, TronNetwork network = TronNetwork.MainNet)
@@ -72,7 +72,7 @@ namespace TronNet
 
             if (_network == TronNetwork.MainNet)
             {
-                _publicAddress = Base58Encoder.Encode(addressChecksum);
+                 _publicAddress = Base58Encoder.Encode(addressChecksum);
             }
             else
             {
